Move meta handling into its own class
