+++ src/Events/Classy/Controller.php
@@ -28,45 +28,6 @@ class Controller extends Controller_Contract {
 	use Supported_Post_Types;
 
 	/**
-	 * The list of event meta keys to be registered.
-	 *
-	 * This list is used to register the post meta fields for the Classy application. The
-	 * key is the meta key, and the value is an array of arguments used in the `register_post_meta`
-	 * function. The `single` key indicates whether the meta field is a single value or an array,
-	 * and the `type` key indicates the type of the value. If no `single` or `type` is provided,
-	 * the default is `single` set to `true` and `type` set to `string`.
-	 *
-	 * In the JS application, these meta fields are defined in a single constants file.
-	 *
-	 * @see src/resources/packages/classy/constants.tsx
-	 * @see self::register_meta_fields()
-	 *
-	 * @var array<array-key, array<string, mixed>>
-	 */
-	private const META = [
-		'_EventAllDay'           => [],
-		'_EventCost'             => [],
-		'_EventCurrency'         => [],
-		'_EventCurrencyPosition' => [],
-		'_EventCurrencySymbol'   => [],
-		'_EventEndDate'          => [],
-		'_EventIsFree'           => [
-			'type' => 'boolean',
-		],
-		'_EventStartDate'        => [],
-		'_EventTimezone'         => [],
-		'_EventURL'              => [],
-		'_EventOrganizerID'      => [
-			'single' => false,
-			'type'   => 'integer',
-		],
-		'_EventVenueID'          => [
-			'single' => false,
-			'type'   => 'integer',
-		],
-	];
-
-	/**
 	 * Registers the hooks and filters for this controller.
 	 *
 	 * @since TBD
@@ -74,7 +35,7 @@ class Controller extends Controller_Contract {
 	 * @return void
 	 */
 	protected function do_register(): void {
-		$this->register_meta_fields();
+		$this->container->register( Meta::class );
 		add_filter( 'tec_classy_post_types', [ $this, 'add_supported_post_types' ] );
 		add_filter( 'tec_classy_localized_data', [ $this, 'filter_data' ] );
 
@@ -94,47 +55,10 @@ class Controller extends Controller_Contract {
 	 * @return void
 	 */
 	public function unregister(): void {
-		$this->unregister_meta_fields();
+		$this->container->get( Meta::class )->unregister();
 		remove_filter( 'tec_classy_post_types', [ $this, 'add_supported_post_types' ] );
 		remove_filter( 'tec_classy_localized_data', [ $this, 'filter_data' ] );
 		remove_action( 'tec_common_assets_loaded', [ $this, 'register_assets' ] );
-	}
-
-	/**
-	 * Registers meta fields for all supported post types.
-	 *
-	 * @since TBD
-	 *
-	 * @return void
-	 */
-	private function register_meta_fields(): void {
-		foreach ( self::META as $meta_key => $args ) {
-			$post_meta_args = [
-				'show_in_rest'  => true,
-				'single'        => $args['single'] ?? true,
-				'type'          => $args['type'] ?? 'string',
-				'auth_callback' => fn() => $this->current_user_can_edit_events(),
-			];
-
-			foreach ( $this->get_supported_post_types() as $post_type ) {
-				register_post_meta( $post_type, $meta_key, $post_meta_args );
-			}
-		}
-	}
-
-	/**
-	 * Unregisters the post meta fields for the plugin.
-	 *
-	 * @since TBD
-	 *
-	 * @return void
-	 */
-	private function unregister_meta_fields(): void {
-		foreach ( self::META as $meta_key => $args ) {
-			foreach ( $this->get_supported_post_types() as $post_type ) {
-				unregister_post_meta( $post_type, $meta_key );
-			}
-		}
 	}
 
 	/**
